comp_util: two functions are added to convert cpu arrays to gpu array.
_convert_to_gpu_array:
Function accepts a python list of inputs and converts numpy to cupy, or xarray array with numpy to xarray array to cupy, and xarray array with numpy dask to xarray array to cupy dask.

_convert_to_cpu_array:
Function accepts a python list of inputs and converts cupy to numpy, or xarray array with cupy to xarray array to numpy, and xarray array with cupy dask
to xarray array to numpy dask.

All the ported functions now are using these two functions.