comp_util: the conversion function are change so to support dask arrays.
crop.py: All functions are posted to GPU.
But since cp.sin and cp.exp doesn't support input type xarray and dask there are some problems. Using GPU for xarray and dask datatypes in these functions is not possible yet. We may need to change the cp.sin and cp.exp to xr.ufunc.sin and xr.ufunc.exp (not tested).
Because of the current issue, some unittests fail.